basic game logic and play 2 turn test.
After dealing the cards, I tried to write a simple game stage that I could control on both sides. I wrote a method called comparetoCards, where I tried to create the background of the game, that is, the mechanism. This method is a method that compares the card written in it according to the number of cards on the board and its value. For now, it is a simple method to test. I play 2 rounds with logic.