about scoreboard
After the game is over and the remaining cards on the board go to the last winner, I wrote a function that navigates the winner's treasury. This function increases the card number by 1 for each expression that does not correspond to null while navigating in the treasure. It adds 1 point per each. Finally, it adds 10 times the number of pishti to this total score and finds the total score. Then, according to each other's card numbers, one gains 3 points more than the other and the total score is determined.If the number of cards is equals, no one gets 3 points.